tried to add support for BIN auctions using /flips, added NullPointerException instead

+++ src/main/java/com/max2341/fruitbin/AuctionInfo.java
@@ -4,7 +4,7 @@ import com.max2341.fruitbin.Utils.Risk;
 
 public class AuctionInfo {
 	
-	public AuctionInfo(AuctionItem auction, int profit, int profitPercent, float price, float lowestBin, int amountListed, Risk risk) {
+	public AuctionInfo(AuctionItem auction, long profit, int profitPercent, float price, float lowestBin, int amountListed, Risk risk) {
 		this.profitPercent = profitPercent;
 		this.auction = auction;
 		this.price = price;
